+++ app.config.ts
@@ -15,7 +15,6 @@ export default defineAppConfig({
     footer: {
       quote: 'Unlock the potential of your web development journey with UnJS - where innovation meets simplicity, and possibilities become limitless.',
       menu: [
-        {},
         {
           title: 'Community',
           items: [
@@ -39,35 +38,35 @@ export default defineAppConfig({
             },
           ],
         },
-        // {
-        //   title: 'Articles',
-        //   items: [
-        //     {
-        //       title: 'Learn',
-        //       url: '/learn',
-        //       rel: null,
-        //       target: null,
-        //     },
-        //     {
-        //       title: 'Build',
-        //       url: '/build',
-        //       rel: null,
-        //       target: null,
-        //     },
-        //     {
-        //       title: 'Explore',
-        //       url: '/explore',
-        //       rel: null,
-        //       target: null,
-        //     },
-        //     // {
-        //     //   title: 'Search',
-        //     //   url: '/search',
-        //     //   rel: null,
-        //     //   target: null,
-        //     // },
-        //   ],
-        // },
+        {
+          title: 'Content',
+          items: [
+          //   {
+          //     title: 'Learn',
+          //     url: '/learn',
+          //     rel: null,
+          //     target: null,
+          //   },
+          //   {
+          //     title: 'Build',
+          //     url: '/build',
+          //     rel: null,
+          //     target: null,
+          //   },
+          //   {
+          //     title: 'Explore',
+          //     url: '/explore',
+          //     rel: null,
+          //     target: null,
+          //   },
+            {
+              title: 'Search',
+              url: '/search',
+              rel: null,
+              target: null,
+            },
+          ],
+        },
         {
           title: 'UnJS',
           items: [
