+++ components/app/search/AppSearchEmpty.vue
@@ -1,0 +1,6 @@
+<template>
+  <div p="y-4" flex="~ col items-center" gap-2>
+    <span i="heroicons-magnifying-glass-20-solid" text="gray-400" w-5 h-5 block />
+    <span text="gray-600"> We couldn't find any items with that term. Please try again. </span>
+  </div>
+</template>
